+++ CHANGES.md
@@ -10,6 +10,8 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+### Changed
+* Update default google-java-format from 1.9 to 1.10.0
 
 ## [2.13.0] - 2021-03-05
 ### Added