+++ CHANGES.md
@@ -10,6 +10,8 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+
+## [2.13.3] - 2021-04-20
 ### Fixed
 * LicenseHeaderStep treats address as copyright year ([#716](https://github.com/diffplug/spotless/issues/716))